requirements.txt
typer
requests
python-dotenv
manim